Improve jsdoc build process.
Lock down JSDoc to version 3.3.2, because 3.3.3 no longer allows ES6. See https://github.com/jsdoc3/jsdoc/issues/1076
This also removes the 'static assets' folder for the doc build, and puts it in a temporary folder. Currently the only 'static asset' is a minified build of knex.js. We don't want this under version control and it does not need to exist outside of the build process. The Bookshelf theme provides all other non-generated assets (images etc).